Springify

+++ registry-core-api/src/main/java/eu/openminted/registry/core/domain/Value.java
@@ -3,12 +3,12 @@ package eu.openminted.registry.core.domain;
 public class Value implements Comparable<Value> {
 	private String value;
 	private String label;
-	private int count;
+	private long count;
 
 	public Value() {
 	}
 
-	public Value(String value, int count) {
+	public Value(String value, long count) {
 		this.value = value;
 		this.count = count;
 	}
@@ -19,10 +19,10 @@ public class Value implements Comparable<Value> {
 	public void setValue(String value) {
 		this.value = value;
 	}
-	public int getCount() {
+	public long getCount() {
 		return count;
 	}
-	public void setCount(int count) {
+	public void setCount(long count) {
 		this.count = count;
 	}
 
@@ -36,6 +36,6 @@ public class Value implements Comparable<Value> {
 
 	@Override
 	public int compareTo(Value o) {
-		return Integer.compare(count, o.count);
+		return Long.compare(count, o.count);
 	}
 }
